Switched the image to a more reliable github stored one, added function (see desc)
added function getClosestPointInMatrix(obj, matrix, value, int)
the obj need to have a obj.x and obj.y contained in the matrix, the matrix is an array of arrays, the value needs to be in the matrix, and the int is the max number of iterations.
This function finds the value specified in the matrix the closest to the specified point, in a spiral.